Implement running pg_basebackup for cloud backups
This commit mainly implements the `_run_pg_basebackup` method of the
`CloudPostgresBackupExecutor` class.

This method is responsible for starting the `pg_basebackup` process.
It spawns the process asynchronously i.e. it does not wait for the process
to complete. This is important because as soon as the process is started
other threads will be spawned to monitor the staging area, pausing/resuming
the `pg_basebackup`, the fetching of ready files, etc.

References: BAR-958

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>