Implement monitoring of the staging area during cloud backups
Implemented the `CloudPostgresBackupExecutor._monitor_staging_area` method.

It is responsible for continuously checking the size of the staging area
directory and pausing/resuming the `pg_basebackup` process so that it
never goes beyond the threshold configured in `cloud_staging_max_size`.

Auxiliary methods, such as `pause`, `resume` and `is_running` were also
added to the `Command` class so that the `pg_basebackup` process
(an instance of `PgBaseBackup`, which inherits from `Command`) can be
managed accordingly.

References: BAR-959

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>