Implement fetching of ready files in the staging area during cloud backups
It mainly implements the `CloudPostgresBackupExecutor._fetch_ready_files`
method.

The method is used to retrieve files that are ready for upload from the
staging directory (essentially, files that are closed). It continuously
scans the staging area for new files that have been fully written and
are safe to upload to cloud storage.

In order to know which files are still open (being written to, i.e. not
ready for upload), it makes use of the `lsof` command. To facilitate
the usage of `lsof`, a new command class `Lsof` was also created.

References: BAR-960

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>