Implement building and uploading of tarballs for cloud backups
Implemented the `CloudPostgresBackupExecutor._upload_ready_files` method.

The method is responsible for fetchign ready files from the ready queue,
building tarballs from such files, and uploading the tarballs to the cloud
storage configured.

The core logic of build and upload is not new. It is the same logic used
by the Barman cloud scripts. This is implemented by resuing the
`CloudUploadController` class from the `barman.cloud` module.

The only adaptation needed in the existing cloud module was to allow the
`CloudTarUploader` class (used by the `CloudUploadController`) to accept
a custom staging directory. The staging directory is where the tarballs
are temporarily stored before being uploaded to the cloud. As this
feature supports a `cloud_staging_directory` configuration, such change
was required.

References: BAR-961

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>